Tried to do parallel processing for while loop
Attempted to make it so one thread did a while loop that depended on a
field and the other thread allowed the user to control the field. Didn't
work since apparently the change is not commuted between threads. Will
have to quit our program's while loop using ctrl + c or a built in timer
of sorts. If you have any ideas, let me know.