Classification Decision Tree
Hopefully works. It saves to a file after each run and you can recall an
old one if you want. Does not predict when "learning", so you will have
to rerun without learning turned on to have it throw predictions. Still
need to incorporate the third electrode and generalize the code
regarding electrode names.