Added user feedback and fixed channel names
User feedback now when making predictions. Variable called "running"
that is set by the user when they 'quit' the learning or predicting,
which stops the runSMILE while loop. Removed the old machine learning
algorithm and saved it to another file with Greg's response block in
case we want to use it ever. Fixed the channel names so that now they
are left, right, and center (R, L, or C when combined with other words).
Now when finding the RQ, we subtract magC from both magR and magL before
dividing L/R.