Reorganized, made improvements, added audio
Folder now is cleaned up. Graphs and output files are in 'bin', matlab
files we don't use much are in 'src'. Max's audio files are in 'audio',
publisher files and pdfs are in 'doc', header and dll files are in
'lib'.

The program now has intro text for the two aspects of the demo and for
the main program as well. EmotivSMILE can incorporate 'indifferent' into
its recordings, though for Greg's demo this won't matter. It is mostly
there because when the panelist does machine learning, they can see it
is a possibility.

Max's audio clips now play when logical. I used the ones titled
"Detecting...", "Happy Sad or Neither", and "Would you agree".